data/user/projects/proj_1774533386979_c8fff7/repo/requirements.txt
# 无人机智能感知与自主导航系统 依赖清单
# Python 3.10 / Ubuntu 22.04 / ROS2 Humble

# 深度学习与模型训练
torch==2.1.0
torchvision==0.16.0
onnx==1.15.0
onnxruntime==1.16.3

# 计算机视觉
opencv-python==4.8.1.78
numpy==1.26.2
scipy==1.11.4
Pillow==10.1.0

# 点云与3D处理
open3d==0.17.0

# 数据处理与可视化
matplotlib==3.8.2
pandas==2.1.3
PyYAML==6.0.1
tqdm==4.66.1

# 传感器通讯与消息队列
pyserial==3.5
paho-mqtt==1.6.1

# 标定与几何
transforms3d==0.4.1
apriltag==0.0.16

# 测试
pytest==7.4.3
pytest-cov==4.1.0
